+++ core/src/main/php/io/streams/Streams.class.php
@@ -211,9 +211,9 @@
       if (!isset(self::$streams[$id])) {
         return FALSE;
       } else if ('r' === $m) {
-        return array('size' => 0);
+        return array('size' => 0, 'mode' => 0100644);
       } else if ('w' === $m) {
-        return array('size' => 0);
+        return array('size' => 0, 'mode' => 0100644);
       }
     }